When removing inactive servers from the database make sure all schedules are updated to a new server. Also make sure that we are checking for valid servers when the task is added to the scheduler queue.
