+++ DNN Platform/Library/Services/Scheduling/Scheduler.cs
@@ -105,6 +105,46 @@ namespace DotNetNuke.Services.Scheduling
                 return scheduleHistoryItem;
             }
 
+            /// <summary>
+            /// Checks the scheduled item if it has servers specified to run on. If it does, then it checks to see if the server is active.
+            /// </summary>
+            /// <param name="scheduleItem">The schedule to check</param>
+            /// <returns>True if nothing happens to the scheduled item. Returns false if the scheduled item was updated.</returns>
+            /// <remarks>
+            /// If the server specified is not active then the scheduled item will be assigned to the most recent active server.
+            /// This assumes that all schedules will run on a single server if none of the servers specified are active.
+            /// </remarks>
+            public static bool ValidateServersAreActiveForScheduledItem(ScheduleItem scheduleItem)
+            {
+                // If this scheduled items runs on all servers, continue.
+                if (string.IsNullOrEmpty(scheduleItem.Servers))
+                {
+                    return true;
+                }
+
+                // Get the individual server names from the scheduled item
+                var servers = scheduleItem.Servers.ToLowerInvariant().Split(new[] { ',' }, StringSplitOptions.RemoveEmptyEntries);
+
+                // Get the enabled servers with recent activity.
+                var enabledServers = ServerController.GetEnabledServersWithActivity();
+
+                // Validate that any server in the scheduled item is in the enabled server list with activity in the last 10 minutes.
+                bool serverHasRecentActivity = enabledServers.Any(i => servers.Contains(i.ServerName.ToLowerInvariant()));
+
+                // If no servers in the scheduled item had recent activity, assign this scheduled item to the server with the most recent activity
+                if (!serverHasRecentActivity)
+                {
+                    // Assign a new server to this task then run it
+                    scheduleItem.Servers = $",{enabledServers.First().ServerName.ToLowerInvariant()},";
+
+                    // Save the scheduled item back to the database.
+                    SchedulingController.UpdateSchedule(scheduleItem);
+                    return false;
+                }
+
+                return true;
+            }
+
             /// <summary>Adds an item to the collection of schedule items in queue.</summary>
             /// <param name="scheduleHistoryItem"></param>
             /// <remarks>Thread Safe.</remarks>
@@ -119,7 +159,8 @@ namespace DotNetNuke.Services.Scheduling
                         {
                             // Do a second check just in case
                             if (!ScheduleQueueContains(scheduleHistoryItem) &&
-                                !IsInProgress(scheduleHistoryItem))
+                                !IsInProgress(scheduleHistoryItem) &&
+                                ValidateServersAreActiveForScheduledItem(scheduleHistoryItem))
                             {
                                 // It is safe for this thread to read or write
                                 // from the shared resource.
