+++ DNN Platform/Library/Services/Scheduling/SchedulingController.cs
@@ -257,5 +257,15 @@ namespace DotNetNuke.Services.Scheduling
 
             return false;
         }
+
+        /// <summary>
+        /// Replaces the old server name, with the new server name on all schedules where the old server name was found.
+        /// </summary>
+        /// <param name="oldServer">The old server to replace</param>
+        /// <param name="newServer">The new server to use</param>
+        internal static void ReplaceServer(ServerInfo oldServer, ServerInfo newServer)
+        {
+            DataProvider.Instance().ReplaceServerOnSchedules(oldServer.ServerName, newServer.ServerName);
+        }
     }
 }
